Move SearchOption base classes to new file
